fix scrolling bug on bird list reordering
Before this fix, the bird list would scroll to a bird component
if there was a given hash in the url everytime the birdlist
was updated. Such as the list being reordered.

This commit adds stae.hasScrolled so that the scroll will only
occur once, after the new birds have been received.