Restructure the component and containers directory
These files were originally divided inlign with Dan Abramov's (of React) view on
presentational and container components. However, this structure was quite
awkward to maintain and even he has moved away from the idea of dividing
components like this.

This change will cause there to only be a components folder with subfolders.
The app component will be at the root of this directory and other components
will be grouped into the subfolders based on features unless they are shared.