Make OptionsBar and BackLink components use opacity on hover
Replacement for the "ugly" text-decoration, to give it a more modern
look and be consistent with the rest of the website. In order to clear
the browser's default of text-decoration on anchors, it was necessary
to set none too all of them. This was added in all.scss and removed
elsewhere to prevent duplication.