Add observed_at to Observation and API controller
When support is added to the frontend, this will allow the user to save
a date that differs from the day they are creating it. Meaning, they
can for example, do it the day after or even make observation records
for dates for the distant past.

There will be a scenario where the user wants to create an observation
but does not know when they observed the bird at. For this, we should
pass 0 to the API so that the record gets saved as nil. (The db only
supports nil/NULL and dates).

The reason we want to pass 0 instead of nil, is so that the API
explicitly has to delare that the date is unknown, preventing accidental
requests without the observed at (nil).