Add component to filter Birds by Order and Family
By using the already written updateFilters action, this component will
call it when either the Order or Family select is changed.

An important part of this component is that when the Order is changed,
it should also change the families options to only include families of
that order.

Default values of the Order and Family select dropdowns will ensure that
the filters are set to null independently, meaning that all birds or all
birds for an order (if only the order is set) are rendered.