Use name attributes in components based on userSettings
The Bird and BirdDetails components will now display the bird names
based on the user's primaryNameLanguage and secondaryNameLanguage,
rather than being hardcoded.

The FamilyAndOrderFilter component / family and order dropdowns will
now display the orders in their English, Swedish or Scientific name
based on the primaryNameLanguage. Note, the 'orders' and 'families' text
will still remain in English as the app is currently in English.

Later on, a settings component will be made so that the user can change
these name language settings/preferences.