Extract Bird fixtures from filter_birds.test to a separate file
This will be useful as we can reuse the fixtures in other tests.

The initialBirds was left in the test file, as each test should be
responsible for only using the fixtures that it needs - having less
fixtures means less code and easier to read, but some tests require more
as they need more diverse data.