Add sorting logic
Note, sorting on 'seen' in ascending order will order birds like this:
1. Seen birds with unknown date
2. Seen birds with known dates (with the earlier dates occuring before)
3. Unseen birds

The reason for putting unknown dates before known dates is because
observers will generally not know the observation dates from the past,
where as known dates will be either recorded after these or made in the
app at the time.

The updateSorting action and BirdList column header component will soon
be added which will allow the user to sort the birds by interaction.