Change Bird scope to exclude non-occurring birds
Some birds have been added to the database that have never been seen in
Sweden. These birds have a population category of 0, and should be
excluded from being used in the app. The best way to do this is through
a default scope so that they are consistently excluded from all SQL
calls.