Change non-occurring birds' population category from nil to zero
Some birds are in the database which do not occur in Sweden. These had
a population category of nil. Because of this, it was preventing a
presence validator. Also, in the future, we will adapt the Bird default
scope to exclude these birds, and in-conjuction with not having a
presence validator, it would be confusing to forget to set the category
and then not be able to find the Bird (due to the default scope).

It is also better that to have to set the population_category to a value
that has a special meaning, rather than it being the default.