Add clickSortingColumn helper
This helper is responsible for determining what the new sorting values
should be by looking at the clicked header and what the sorting values
are currently.

Essentially, it boils down to, if the same header is clicked, increment
the ordering ASC -> DESC -> default (null header column). If a new
header is clicked, it should be set to that column and ASC ordering.