Add sortedBirds to state
Added to the state and the BirdList component will now use this data. As
of this commit, the user cannot modify the sorting data, but the
filteredBirds will still go through the sortBirds helper. As the sorting
state column is null by default, the sortedBirds will be the same as the
filteredBirds data. Meaning, this change has no visual effect yet.