Use Redux state for SeenBirdsFilter component
By storing the SeenBirdsFilter value in Redux and using it as the source of
truth for the component, it will allow us to control it by other actions
or components.

This is needed for the upcoming reset filters button.

Original code by Sara Lotfi and commits reorganised by Ryan Woods.

Co-authored-by: Sara Lotfi <lotfi.sara@gmail.com>