Fix search filter bug when no birds were found
The filterBirds helper was treating an empty searchScope (result of
the search query), as no search has been done, so ignore this filter and
only filter by other filters. The problem was, that it meant that if
someone typed a query that resulted in no birds, it would then ignore
the searchScope, resulting in birds still being shown.

To prevent this, we'll return an empty array when a search query is set
and the search result (searchScope) was an empty array.

Note, we could do this check in the iterator, however, this is
unnecessary and prevents a possible 500 interator checks.