Fix search returning no birds when search query is one character
Because of the type of search the backend is using (trigram), a result
of no birds are returned when the search query is only one character in
length.

Generally, filterBirds wants to iterate over each bird and then check
if they are included in the search result (searchScope). However, there
is a guard clause at the start of the method, to return an empty array
when the result is nothing, to prevent unnecessary iterations. Before,
this guard also checked that the query was at least one character in
length, but to fix this it will be changed to at least two characters.