Refactor ObservationsControllerTest
The goal here was to reduce the number of instance variables. Having
many but similar variables makes the code more difficult to read as the
reader needs to understand how the variables relate to each other. For
example, before, the following existed in the setup:
- @today = Date.today
- @observed_at = @today.to_s

During one test, both of these variables may be used, one for the
arrange, and one for the assertion. Making it more difficult to read
rather than just using one variable.