Add ObservationsController#update endpoint
This endpoint allows for editing an observation.

Note, a different observations route resource was created as an id must
be given for PATCH requests, and we do not want to force the user to
give both the scientific_name of the bird via params[:bird_id] and the
id of the observation via params[:id], as the observations id is not
available via the frontend. Instead, params[:id] will be used for the
bird scientific_name, allowing for finding of the observation as users
can only have one observation for each bird.