Change ObservationModal to send 0 based on dateUnknown
It's more reliable to look at the dateUnknown checkbox value on whether
we should send 0 or not (0 means the date is unknown), instead of an
empty string on observedAt date. The observedAt date is currently set
to an empty string when the dateUnknown is checked, and therefore it is
a side-effect. If we ever changed this behaviour, then the inferring of
0 would have not worked correctly.