Use link not link_directory in assets manifest
Linking the whole directory was causing the following error:

```
ActionView::Template::Error (Asset 'application.css' was not declared to be precompiled in production.
Declare links to your assets in 'app/assets/config/manifest.js'.

  //= link application.css
and restart your server):
```

Instead, it is enough to just link the main application.css as this
imports the other css files necessary (which also import files and so
on).

References:
- https://github.com/rails/sprockets-rails/issues/458#issuecomment-620004773
- https://github.com/rails/sprockets/issues/161#issuecomment-159023538