Improve serialized array test assertion fail messages
Before, it was very difficult to read the fail messages regarding
serialized arrays.

1. Because you have to locate where the expected item is in the array
2. It might not even be in there
3. If it is, it's difficult to see clearly what attributes are different

Now, with this helper, you give the actual, expected and a useful attribute
name, then it will check if the item is even in the array, and then if
it is, it will assert if is what you expect it should be.

This produces easy to read error messages for both key areas:
- element inclusion; and
- element equality.