Move modal-container out of page-container
There's some CSS on the page-container [1], to cause its contents to
grow in size to ensure that the footer is always the bottom of the
page/not float in the middle. When the content-container was too short
in height to fillup the viewport, it triggered the CSS to increase the
height of the modal-container, which meant empty space above the
filters when no Birds were rendered.

[1] https://github.com/RyanofWoods/rails-react-swedish-birds/blob/5119a02ef182d02637a4587ba4988a2457e321f2/app/assets/stylesheets/components/_container.scss#L4